\name{generalhoslem-package}
\alias{generalhoslem-package}
\alias{generalhoslem}
\docType{package}
\title{
\packageTitle{generalhoslem}
}
\description{
\packageDescription{generalhoslem}
}
\details{
\packageDESCRIPTION{generalhoslem}
\packageIndices{generalhoslem}
}
\author{
\packageAuthor{generalhoslem}
Maintainer: \packageMaintainer{generalhoslem}
}
\references{
\itemize{
\item Fagerland MW, Hosmer DW, Bofin AM. Multinomial goodness-of-fit tests for logistic regression models. Statistics in Medicine 2008;27(21):4238-53.
\item Fagerland MW, Hosmer DW. A goodness-of-fit test for the proportional odds regression model. Statistics in Medicine 2013;32:2235-2249.
\item Fagerland MW, Hosmer DW. Tests for goodness of fit in ordinal logistic regression models. Journal of Statistical Computation and Simulation 2016. DOI: 10.1080/00949655.2016.1156682.
\item Hosmer DW, Lemeshow S, Sturdivant RX. Applied Logistic Regression, 3rd Edition. 2013. New York, USA: John Wiley and Sons.
\item Lipsitz SR, Fitzmaurice GM, Molenberghs G. Goodness-of-Fit Tests for Ordinal Response Regression Models. Journal of the Royal Statistical Society 1996;45(2):175-190.
\item Pulkstenis E, Robinson TJ. Goodness-of-fit tests for ordinal response regression models. Statistics in Medicine 2004;23:999-1014.
}
}
\keyword{ package }